Refactor Playwright interface to run page.evaluate
