+++ lib/browser-interface-iframe.js
@@ -80,7 +80,9 @@ class BrowserInterfaceIframe extends BrowserInterface {
 			await this.resize( viewport );
 		}
 
-		return method( this.iframe.contentWindow, ...args );
+		// The inner window in the iframe is separate from the main window object.
+		// Pass the iframe window object to the evaluating method.
+		return method( { innerWindow: this.iframe.contentWindow, args } );
 	}
 
 	addGetParameters( rawUrl ) {
