+++ lib/browser-interface.js
@@ -42,8 +42,9 @@ class BrowserInterface {
 		);
 	}
 
-	static innerGetCssIncludes( window ) {
-		return [ ...window.document.getElementsByTagName( 'link' ) ]
+	static innerGetCssIncludes( { innerWindow } ) {
+		innerWindow = null === innerWindow ? window : innerWindow;
+		return [ ...innerWindow.document.getElementsByTagName( 'link' ) ]
 			.filter( ( link ) => link.rel === 'stylesheet' )
 			.reduce( ( set, link ) => {
 				set[ link.href ] = {
@@ -59,10 +60,13 @@ class BrowserInterface {
 	 * for easy querySelector calling (values), return an array of selectors which match
 	 * _any_ element on the page.
 	 *
-	 * @param {Window} innerWindow - Window inside the browser interface.
-	 * @param {Object} selectors - Map containing selectors (object keys), and simplified versions for easy matching (values).
+	 * @param {Object} wrappedArgs
+	 * @param {Window} wrappedArgs.innerWindow - Window inside the browser interface.
+	 * @param {Object[]} wrappedArgs.args - Array of arguments.
+	 *        {Object} wrappedArgs.args[selectors] - Map containing selectors (object keys), and simplified versions for easy matching (values).
 	 */
-	static innerFindMatchingSelectors( innerWindow, selectors ) {
+	static innerFindMatchingSelectors( { innerWindow, args: [ selectors ] } ) {
+		innerWindow = null === innerWindow ? window : innerWindow;
 		return Object.keys( selectors ).filter( ( selector ) => {
 			try {
 				return !! innerWindow.document.querySelector(
@@ -80,21 +84,23 @@ class BrowserInterface {
 	 * for easy querySelector calling (values), return an array of selectors which match
 	 * any above-the-fold element on the page.
 	 *
-	 * @param {Window} innerWindow - Window inside the browser interface.
-	 * @param {Object} selectors - Map containing selectors (object keys), and simplified versions for easy matching (values).
-	 * @param {string[]} pageSelectors - String array containing selectors that appear anywhere on this page (as returned by innerFindMatchingSelectors) - should be a subset of keys in selectors.
+	 * @param {Object} wrappedArgs
+	 * @param {Window} wrappedArgs.innerWindow - Window inside the browser interface.
+	 * @param {Object[]} wrappedArgs.args - Array of arguments.
+	 *        {Object} wrappedArgs.args[selectors] - Map containing selectors (object keys), and simplified versions for easy matching (values).
+	 *        {string[]} wrappedArgs.args[pageSelectors] - String array containing selectors that appear anywhere on this page (as returned by innerFindMatchingSelectors) - should be a subset of keys in selectors.
 	 */
-	static innerFindAboveFoldSelectors(
+	static innerFindAboveFoldSelectors( {
 		innerWindow,
-		selectors,
-		pageSelectors
-	) {
+		args: [ selectors, pageSelectors ],
+	} ) {
 		/**
 		 * Inner helper function used inside browser / iframe to check if the given
 		 * element is "above the fold".
 		 *
 		 * @param {HTMLElement} element - Element to check.
 		 */
+		innerWindow = null === innerWindow ? window : innerWindow;
 		const isAboveFold = ( element ) => {
 			const originalClearStyle = element.style.clear || '';
 			element.style.clear = 'none';
